+++ src/de/litigame/Program.java
@@ -2,6 +2,7 @@ package de.litigame;
 
 import de.gurkenlabs.litiengine.Game;
 import de.gurkenlabs.litiengine.resources.Resources;
+import de.litigame.gui.IngameScreen;
 
 public class Program {
